analyzer: split simplification into its own directory.
