+++ analyzer/expressions/binary.cpp
@@ -51,34 +51,6 @@ inline namespace _v1
         os << in << "}\n";
     }
 
-    std::unique_ptr<expression> binary_expression::_clone_expr_with_replacement(replacements & repl) const
-    {
-        auto ret = std::unique_ptr<binary_expression>(new binary_expression(*this));
-
-        ret->_lhs = _lhs->clone_expr_with_replacement(repl);
-        ret->_rhs = _rhs->clone_expr_with_replacement(repl);
-
-        return ret;
-    }
-
-    future<expression *> binary_expression::_simplify_expr(simplification_context & ctx)
-    {
-        return when_all(_lhs->simplify_expr(ctx), _rhs->simplify_expr(ctx))
-            .then([&](auto && simplified) {
-                replace_uptr(_lhs, get<0>(simplified), ctx);
-                replace_uptr(_rhs, get<1>(simplified), ctx);
-                return _overload->simplify(ctx, { _lhs->get_variable(), _rhs->get_variable() });
-            })
-            .then([&](auto && simplified) -> expression * {
-                if (simplified)
-                {
-                    return simplified;
-                }
-
-                return this;
-            });
-    }
-
     statement_ir binary_expression::_codegen_ir(ir_generation_context & ctx) const
     {
         auto lhs_instructions = _lhs->codegen_ir(ctx);
