+++ analyzer/simplification/expressions/expression_list.cpp
@@ -32,33 +32,22 @@ namespace reaver::vapor::analyzer
 {
 inline namespace _v1
 {
-    void expression_list::print(std::ostream & os, std::size_t indent) const
+    future<expression *> expression_list::_simplify_expr(simplification_context & ctx)
     {
-        auto in = std::string(indent, ' ');
-        os << in << "expression list at " << range << '\n';
-        os << in << "type: " << value.back()->get_type()->explain() << '\n';
-        fmap(value, [&](auto && expr) {
-            os << in << "{\n";
-            expr->print(os, indent + 4);
-            os << in << "}\n";
-
-            return unit{};
+        return when_all(fmap(value, [&](auto && expr) { return expr->simplify_expr(ctx); })).then([&](auto && simplified) -> expression * {
+            replace_uptrs(value, simplified, ctx);
+            assert(0);
+            return this;
         });
     }
 
-    std::unique_ptr<expression> preanalyze_expression_list(const parser::expression_list & expr, scope * lex_scope)
+    std::unique_ptr<expression> expression_list::_clone_expr_with_replacement(replacements & repl) const
     {
-        if (expr.expressions.size() == 1)
-        {
-            return preanalyze_expression(expr.expressions.front(), lex_scope);
-        }
-
         auto ret = std::make_unique<expression_list>();
-        ret->value.reserve(expr.expressions.size());
-        std::transform(expr.expressions.begin(), expr.expressions.end(), std::back_inserter(ret->value), [&](auto && expr) {
-            return preanalyze_expression(expr, lex_scope);
-        });
-        ret->range = expr.range;
+        ret->range = range;
+
+        ret->value = fmap(value, [&](auto && expr) { return expr->clone_expr_with_replacement(repl); });
+
         return ret;
     }
 }
